scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.28.0
